+++ ch05/instructions/loads/lload.go
@@ -6,7 +6,7 @@ import (
 )
 
 // Load Long variable from LocalVar table and push to opstack
-func _lload(frame *rtda.StackFrame, index uint8) {
+func _lload(frame *rtda.StackFrame, index uint) {
 	val := frame.LocalVars().GetLong(index)
 	frame.OperandStack().PushLong(val)
 }
